# my_basics/openApi.yml
openapi: 3.0.3
info:
  title: EazyShop Products APIs Definition
  description: | 
    # About us
    `Using` this OpenAPI specification, any consumer
    can understand the _APIs_ that are exposed by **EasyShop**
    #Categories supported
    - Mobile
      - Apple
      - Samsung
      - OnePlus
    - Laptops
    - TVs
    - HeadPhones
    ```
    test highlight 
    multiline
    ```
  termsOfService: https:eazyshop.com/terms/
  x-custom-info:
    comment: Some comments
    developers:
      - name: John Doe
        email: john@doe.com
      - name: Jane Doe
        email: jane@doe.com
  contact:
    name: API Support
    url: https:eazyshop.com/support
    email: support@eazyshop.com
  license: 
    name: EasyShop 2.0
    url: https:eazyshop.com/licence/LICENCE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://development.eazyshop-server.com/v1
    description: Development server
  - url: https://staging.eazyshop-server.com/v1
    description: Staging server
  - url: https://api.eazyshop-server.com/v1
    description: Production server
  - url: https://{username}.server.com:{port}/{version}
    variables:
      username:
        default: demo
        description: This value is assigned by the service provider.
      port:
        enum:
          - "8443"
          - "443"
        default: "8443"
      version:
        default: v1
...
security:
  - BasicAuth: []
  - BearerAuth: []
  - ApiKeyAuth: []
  - oAuth2AuthCode:
    - read
    - write
    - admin
paths: 
...
  /products:
    get:
      tags:
        - Products
      summary: List all Products
      description: |
        Returns the list of categories supported by EasyShop
      operationId: getAllProducts
      parameters:
        - name: categoryId
          in: query
          schema: 
            type: integer
          required: false
          example: 105
      responses:
        '200':
          description: A list of products
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: "#/components/schemas/Product"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Return Product etail
      description: |
        Returns the detail of product
      operationId: getProductDetails
      parameters:
        - name: productId
          in: path
          schema: 
            type: integer
          required: true
          example: 105
      responses:
        '200':
          description: Detail of product
          content: 
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Product"
...
components:
...
  schemas:
...
    Product:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
        categoryName:
          type: string
          nullable: true
        releaseDate:
          type: string
          format: date
          readOnly: true
        quantity:
          type: integer
          description: |
            ***Quantity*** represents `stock` value when 
            this field is send to EasyShop <u>xxx </ul>
...
  responses:
    InternalServerError:
      description: |
        An error occured while processing the request.
        Please contact support@eazyshop.com
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:  # custom security here in header      
      type: apiKey
      in: header
      name: X-API-KEY
    oAuth2AuthCode:
      type: oauth2
      flows: 
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/api/oauth.access
          scopes:
            read: Read only access
            write: Read Write access
            admin: Admin access